API reference

Keys

TFHE.SchemeParameters — Type

TFHE scheme parameters (single- or multi- party).

TFHE.tfhe_parameters — Function

tfhe_parameters(; tlwe_mask_size::Int=1)

Creates a single-party SchemeParameters object to pass to SecretKey.

TFHE.SecretKey — Type

SecretKey(rng::AbstractRNG, params::SchemeParameters)

A TFHE secret key, used for encryption/decryption. Currently the only official way to get an object to pass to params is from tfhe_parameters.

TFHE.CloudKey — Type

CloudKey(rng::AbstractRNG, secret_key::SecretKey)

A TFHE cloud (public) key, used for secure computations by a third party.

TFHE.make_key_pair — Function

make_key_pair(rng::AbstractRNG, params::Union{Nothing, SchemeParameters}=nothing)

Creates a pair of SecretKey and a corresponding CloudKey. If params is nothing, the default return value of tfhe_parameters is used.

Encryption/decryption

TFHE.encrypt — Function

encrypt(rng::AbstractRNG, key::SecretKey, message::Bool)

Encrypts a plaintext bit. Returns a LweSample object.

TFHE.decrypt — Function

decrypt(key::SecretKey, sample::LweSample)

Decrypts an encrypted bit. Returns a boolean.

TFHE.LweSample — Type

A structure representing an encrypted bit.

Logical gates

TFHE.gate_nand — Function

gate_nand(ck::CloudKey, x::LweSample, y::LweSample)

Applies the NAND gate to encrypted bits x and y. Returns a LweSample object.

TFHE.gate_or — Function

gate_or(ck::CloudKey, x::LweSample, y::LweSample)

Applies the OR gate to encrypted bits x and y. Returns a LweSample object.

TFHE.gate_and — Function

gate_and(ck::CloudKey, x::LweSample, y::LweSample)

Applies the AND gate to encrypted bits x and y. Returns a LweSample object.

TFHE.gate_xor — Function

gate_xor(ck::CloudKey, x::LweSample, y::LweSample)

Applies the XOR gate to encrypted bits x and y. Returns a LweSample object.

TFHE.gate_xnor — Function

gate_xnor(ck::CloudKey, x::LweSample, y::LweSample)

Applies the XNOR gate to encrypted bits x and y. Returns a LweSample object.

TFHE.gate_not — Function

gate_not(ck::CloudKey, x::LweSample)

Applies the NOT gate to an encrypted bit x. Returns a LweSample object. Does not need to be bootstrapped.

TFHE.gate_constant — Function

gate_constant(ck::CloudKey, value::Bool)

Returns a LweSample object representing the plaintext value.

Note

The returned object is suitable to use with other gates, but it is not encrypted.

TFHE.gate_nor — Function

gate_nor(ck::CloudKey, x::LweSample, y::LweSample)

Applies the NOR gate to encrypted bits x and y. Returns a LweSample object.

TFHE.gate_andny — Function

gate_andny(ck::CloudKey, x::LweSample, y::LweSample)

Applies the ANDNY (ANDNY(x, y) == AND(NOT(x), y)) gate to encrypted bits x and y. Returns a LweSample object.

TFHE.gate_andyn — Function

gate_andyn(ck::CloudKey, x::LweSample, y::LweSample)

Applies the ANDYN (ANDYN(x, y) == AND(x, NOT(y))) gate to encrypted bits x and y. Returns a LweSample object.

TFHE.gate_orny — Function

gate_orny(ck::CloudKey, x::LweSample, y::LweSample)

Applies the ORNY (ORNY(x, y) == OR(NOT(x), y)) gate to encrypted bits x and y. Returns a LweSample object.

TFHE.gate_oryn — Function

gate_oryn(ck::CloudKey, x::LweSample, y::LweSample)

Applies the ORYN (ORYN(x, y) == OR(x, NOT(y))) gate to encrypted bits x and y. Returns a LweSample object.

TFHE.gate_mux — Function

gate_mux(ck::CloudKey, x::LweSample, y::LweSample, z::LweSample)

Applies the MUX (MUX(x, y, z) == x ? y : z == OR(AND(x, y), AND(NOT(x), z))) gate to encrypted bits x, y and z. Returns a LweSample object.

Multi-key TFHE (experimental)

TFHE.mktfhe_parameters_2party — Constant

Multi-key TFHE parameters for 2 parties (a SchemeParameters object).

TFHE.mktfhe_parameters_4party — Constant

Multi-key TFHE parameters for 4 parties (a SchemeParameters object).

TFHE.mktfhe_parameters_8party — Constant

Multi-key TFHE parameters for 8 parties (a SchemeParameters object).

TFHE.SharedKey — Type

SharedKey(rng::AbstractRNG, params::SchemeParameters)

A shared key created by the server. params is one of mktfhe_parameters_2party, mktfhe_parameters_4party, mktfhe_parameters_8party.

TFHE.CloudKeyPart — Type

CloudKeyPart(rng, secret_key::SecretKey, shared_key::SharedKey)

A part of the cloud (computation) key generated independently by each party (since it involves their secret keys). The secret_key is a SecretKey object created with the same parameter set as the SharedKey object.

TFHE.MKCloudKey — Type

MKCloudKey(ck_parts::Array{CloudKeyPart, 1})

A full cloud key generated on the server out of parties' cloud key parts.

TFHE.mk_encrypt — Function

mk_encrypt(rng, secret_keys::Array{SecretKey, 1}, message::Bool)

Encrypts a plaintext bit using parties' secret keys. Returns a MKLweSample object.

TFHE.mk_decrypt — Function

mk_decrypt(secret_keys::Array{SecretKey, 1}, sample::MKLweSample)

Decrypts an encrypted bit using parties' secret keys. Returns a boolean.

TFHE.MKLweSample — Type

A structure representing an encrypted bit in multi-key TFHE.

TFHE.mk_gate_nand — Function

mk_gate_nand(ck::MKCloudKey, x::MKLweSample, y::MKLweSample)

Applies the NAND gate to encrypted bits x and y. Returns a MKLweSample object.